Updated numerics so that they actually work
Provides the foundation for the implementation of the full numeric tower
of sub-types. Each type of Complex, Real, Rational, and Integer is
implemented as its own class with overloaded operators declared as
friend functions. Implicit conversion operators are used to allow for
interoperability between the subtypes in a Mathematically sound fashion

Propsoses the use of a Tagged Union "Number" class that will wrap all of
the sub-types and allow for singular implementations of Scheme's other
built-in numeric procedures i.e. anything that can't be implemented as a
simple operator like +, -, *, >=, <=, etc.

An ad-hoc approach has been used in this prototype, primarily because
attempts at a purely Object Oriented solution via inheritance proved
futile do to the necessity of utilizing downcasts/upcasts between
base/derived class pointers/references to get the type interop to work